ecommerce_api/requirements.txt
asgiref==3.2.10
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.6.0
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
Markdown==3.3.3
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.5
shortuuid==1.0.1
sqlparse==0.4.1
whitenoise==5.2.0
